Updating User Creation, SidHistory Injection and added test drive script
My first push since i released it, sorry for the delay.  THis has an update to the create user portion of the script.  In tests the a single user create function ran in 300 ms instead of 1100 ms.  Hopefully the rest of you see the speed increase.  Sidhistory injection listed as 'bonus content'.  Added test drive, so you can powershell to the badblood created system from a remote machine to run queries on it.